Add info to examples/README.rst & convert to Markdown
The `examples/README.rst` file was rather sparse. In order to improve
visibility of the many Cirq examples we provide, I added summaries of
each example file, as well as some additional general info.

In addition, because in past discussions we agreed that readme files
that are not associated with setup.py files could be changed to
Markdown format, this commit also changes the readme file to Markdown
format.